Move resource services to tRPC errors
