+++ src/server/orpc/routers/index.ts
@@ -1,5 +1,1 @@
-import { resourceRouter } from "@server/orpc/routers/resource";
-
-export const appRouter = {
-  resource: resourceRouter,
-};
+export const appRouter = {};
