+++ src/server/services/resource/resourceMetadata.ts
