+++ src/server/utils/urls/__tests__/validateUrl.test.ts
@@ -1,4 +1,3 @@
-import { ORPCError } from "@orpc/client";
 import { TRPCError } from "@trpc/server";
 
 import {
@@ -12,36 +11,6 @@ import {
   ERROR_URL_MALFORMED,
   validateUrl,
 } from "@server/utils/urls/validateUrl";
-
-jest.mock("@orpc/client", () => {
-  type ORPCErrorOptions = {
-    message?: string;
-    cause?: unknown;
-    status?: number;
-    data?: unknown;
-    defined?: unknown;
-  };
-
-  class ORPCError extends Error {
-    public readonly code: string;
-    public readonly status?: number;
-    public readonly data?: unknown;
-    public readonly defined?: unknown;
-
-    constructor(code: string, ...rest: [ORPCErrorOptions] | []) {
-      const opts: ORPCErrorOptions = rest[0] ?? {};
-      super(opts.message ?? `Mock ORPCError with code ${code}`);
-      this.name = "ORPCError";
-      this.code = code;
-      this.cause = opts.cause;
-      this.status = opts.status;
-      this.data = opts.data;
-      this.defined = opts.defined;
-    }
-  }
-
-  return { __esModule: true, ORPCError };
-});
 
 describe("validateUrl", () => {
   describe("rejects", () => {
@@ -76,7 +45,7 @@ describe("validateUrl", () => {
       expect(() => validateUrl("data:text/html;base64,AAAA")).toThrow(
         ERROR_INVALID_SCHEME,
       );
-      expect(() => validateUrl("ftp://example.com")).toThrow(ORPCError);
+      expect(() => validateUrl("ftp://example.com")).toThrow(TRPCError);
     });
 
     it("username/password present", () => {
@@ -97,7 +66,7 @@ describe("validateUrl", () => {
     });
 
     it("port present (even default port)", () => {
-      expect(() => validateUrl("https://example.com:444/")).toThrow(ORPCError);
+      expect(() => validateUrl("https://example.com:444/")).toThrow(TRPCError);
 
       // WHATWG URL deop default ports when parsing/serializing so if :443, url.port === ""
       expect(() => validateUrl("https://example.com:443/path")).not.toThrow(
